ssz: fixes error in list max size resolving type

+++ ssz/src/main/java/org/ethereum/beacon/ssz/access/SSZField.java
@@ -66,7 +66,7 @@ public class SSZField {
     }
 
     @Override
-    public int maxSize() {
+    public long maxSize() {
       return VARIABLE_SIZE;
     }
